requirements.txt
altair==5.4.1
altgraph==0.17.4
anyio==4.4.0
appdirs==1.4.4
argcomplete==3.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.16.0
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boto3==1.35.10
botocore==1.35.10
Brlapi==0.8.5
Brotli==1.1.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2023.11.17
cffi==1.17.0
chardet==5.2.0
click==8.1.6
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
contourpy==1.0.7
copulas==0.11.1
cryptography==41.0.7
cssselect==1.2.0
ctgan==0.10.1
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
debugpy==1.8.5
decorator==5.1.1
deepecho==0.6.0
defer==1.0.6
defusedxml==0.7.1
dell-recovery==0.0.0
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
duplicity==2.1.4
et-xmlfile==1.1.0
executing==2.1.0
Faker==28.1.0
fasteners==0.18
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.1.0
fonttools==4.46.0
fqdn==1.5.1
freetype-py==2.4.0
fs==2.4.16
fsspec==2024.6.1
geojson-rewind==1.1.0
geomet==1.1.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ads==25.1.0
google-api-core==2.22.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
graphviz==0.20.3
grpcio==1.67.1
grpcio-status==1.67.1
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.4
httpx==0.27.2
idna==3.6
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.2.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==24.3.1
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==10.2.0
mpmath==0.0.0
narwhals==1.13.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==2.12.0
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
plotly==5.24.0
pluggy==1.4.0
progressbar==2.5
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydeck==0.9.1
pygbif==0.6.4
Pygments==2.17.2
PyGObject==3.48.2
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.4
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2024.1
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
pyzmq==26.2.0
rdt==1.12.3
referencing==0.35.1
reportlab==4.1.0
requests==2.31.0
requests-cache==1.2.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rlPyCairo==0.3.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.1
sdmetrics==0.15.1
sdv==1.16.1
seaborn==0.13.2
SecretStorage==3.3.3
selenium==4.26.1
Send2Trash==1.8.3
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.7.2
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.40.0
sympy==1.12
systemd-python==235
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
triton==3.0.0
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
vboxapi==1
wadllib==1.3.6
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.42.0
widgetsnbextension==4.0.13
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
